PR review feedback addressed
1. Added space after breaks
2. Seperating container and actual visual component
3. Removing nbsp
4. Adding package-locks to gitignore
