+++ packages/app/main/src/state/actions/ngrokTunnelActions.ts
@@ -56,9 +56,9 @@ export interface TunnelError {
   errorMessage: string;
 }
 
-export interface TunnelStatusAndTs {
+export interface TunnelStatusAndTimestamp {
   status: TunnelStatus;
-  ts: string;
+  timestamp: string;
 }
 
 export interface NgrokTunnelAction<T> extends Action {
@@ -66,7 +66,7 @@ export interface NgrokTunnelAction<T> extends Action {
   payload: T;
 }
 
-export type NgrokTunnelPayloadTypes = TunnelError | TunnelInfo | TunnelStatusAndTs;
+export type NgrokTunnelPayloadTypes = TunnelError | TunnelInfo | TunnelStatusAndTimestamp;
 
 export function updateNewTunnelInfo(payload: TunnelInfo): NgrokTunnelAction<TunnelInfo> {
   return {
@@ -75,12 +75,12 @@ export function updateNewTunnelInfo(payload: TunnelInfo): NgrokTunnelAction<Tunn
   };
 }
 
-export function updateTunnelStatus(tunnelStatus: TunnelStatus): NgrokTunnelAction<TunnelStatusAndTs> {
+export function updateTunnelStatus(tunnelStatus: TunnelStatus): NgrokTunnelAction<TunnelStatusAndTimestamp> {
   return {
     type: NgrokTunnelActions.setStatus,
     payload: {
       status: tunnelStatus,
-      ts: new Date().toLocaleString(),
+      timestamp: new Date().toLocaleString(),
     },
   };
 }
